refactor: auth 패키지와 member 패키지에 대한 의존 제거
