+++ src/main/java/nextstep/auth/AuthConfig.java
@@ -1,28 +1,25 @@
 package nextstep.auth;
 
+import lombok.RequiredArgsConstructor;
 import nextstep.auth.authentication.UsernamePasswordAuthFilter;
 import nextstep.auth.authorization.AuthenticationPrincipalArgumentResolver;
 import nextstep.auth.context.SecurityContextPersistenceFilter;
 import nextstep.auth.authentication.BasicAuthFilter;
 import nextstep.auth.authentication.BearerTokenAuthFilter;
+import nextstep.auth.service.LoginMemberService;
 import nextstep.auth.token.TokenAuthInterceptor;
 import nextstep.auth.token.JwtTokenProvider;
-import nextstep.member.application.LoginMemberService;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import java.util.List;
 
+@RequiredArgsConstructor
 @Configuration
 public class AuthConfig implements WebMvcConfigurer {
-    private LoginMemberService loginMemberService;
-    private JwtTokenProvider jwtTokenProvider;
-
-    public AuthConfig(LoginMemberService loginMemberService, JwtTokenProvider jwtTokenProvider) {
-        this.loginMemberService = loginMemberService;
-        this.jwtTokenProvider = jwtTokenProvider;
-    }
+    private final LoginMemberService loginMemberService;
+    private final JwtTokenProvider jwtTokenProvider;
 
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
