+++ src/main/java/nextstep/auth/service/LoginMemberService.java
@@ -1,0 +1,7 @@
+package nextstep.auth.service;
+
+import nextstep.member.domain.LoginMember;
+
+public interface LoginMemberService {
+    LoginMember loadUserByUsername(String email);
+}
